// src/statistics/enums/place-tag-code.enum.ts
export enum PlaceTagCode {
  HighPreference = 'HIGH_PREFERENCE',
  MeetingPreferred = 'MEETING_PREFERRED',
  WeekendPreferred = 'WEEKEND_PREFERRED',
  WeekdayPreferred = 'WEEKDAY_PREFERRED',
  FrequentlySelected = 'FREQUENTLY_SELECTED',
  CategoryPopular = 'CATEGORY_POPULAR',
  SafeChoice = 'SAFE_CHOICE',
}